requirements.txt
fastapi
torch
clip-anytorch
starlette
ultralytics
pillow
numpy
opencv-python
soundfile
kokoro
uvicorn
